Further updates per code review

+++ warehouse/query-core/src/main/java/datawave/query/util/ssdeep/NGramScoreTuple.java
@@ -21,7 +21,7 @@ public class NGramScoreTuple implements Serializable, Comparable<NGramScoreTuple
         return baseScore;
     }
 
-    public float getWeightedScore() {
+    public int getWeightedScore() {
         return weightedScore;
     }
